Make `pingpong.jl` similar to Intel MPI PingPong benchmark
The difference between `MPI.jl` and the benchmarks reported by Riken was due to
the fact we were measuring times and throughput in a different way.  The new
version of the `pingpong.jl` is brought (hopefully) closer to what the Intel MPI
PingPong benchmark (used by Riken) does, see
<https://github.com/intel/mpi-benchmarks/blob/8ba5d968272b6e7b384f91b6597d1c4590faf3db/src_c/P2P/imb_p2p_pingpong.c>
for reference.